+++ ConfigWidget.lua
@@ -1,1 +1,0 @@
-Scorpio "AshToAsh.BlizzardSkin.Template.Config" ""